--- a/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/dashboard/ExtractedAreaEntity.kt
+++ b/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/dashboard/ExtractedAreaEntity.kt
@@ -2,8 +2,8 @@ package fr.gouv.cacem.monitorenv.domain.entities.dashboard
 
 data class ExtractedAreaEntity(
     val inseeCode: String?,
-    val reportings: List<Int>,
-    val regulatoryAreas: List<Int>,
-    val amps: List<Int>,
-    val vigilanceAreas: List<Int>,
+    val reportingIds: List<Int>,
+    val regulatoryAreaIds: List<Int>,
+    val ampIds: List<Int>,
+    val vigilanceAreaIds: List<Int>,
 )