--- a/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/repositories/IAMPRepository.kt
+++ b/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/repositories/IAMPRepository.kt
@@ -8,5 +8,5 @@ interface IAMPRepository {
 
     fun count(): Long
 
-    fun findAllIdByGeometry(geometry: Geometry): List<Int>
+    fun findAllIdsByGeometry(geometry: Geometry): List<Int>
 }