--- a/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/repositories/IVigilanceAreaRepository.kt
+++ b/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/repositories/IVigilanceAreaRepository.kt
@@ -14,5 +14,5 @@ interface IVigilanceAreaRepository {
 
     fun archiveOutdatedVigilanceAreas(): Int
 
-    fun findAllIdByGeometry(geometry: Geometry): List<Int>
+    fun findAllIdsByGeometry(geometry: Geometry): List<Int>
 }