--- a/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/adapters/bff/outputs/dashboards/ExtractedAreaDataOutput.kt
+++ b/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/adapters/bff/outputs/dashboards/ExtractedAreaDataOutput.kt
@@ -1,10 +1,6 @@
 package fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs.dashboards
 
 import fr.gouv.cacem.monitorenv.domain.entities.dashboard.ExtractedAreaEntity
-import fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs.AMPDataOutput.Companion.fromAMPEntity
-import fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs.RegulatoryAreaDataOutput.Companion.fromRegulatoryAreaEntity
-import fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs.reportings.ReportingDataOutput
-import fr.gouv.cacem.monitorenv.infrastructure.api.adapters.bff.outputs.vigilanceArea.VigilanceAreasDataOutput
 
 class ExtractedAreaDataOutput(
     val inseeCode: String?,
@@ -17,15 +13,10 @@ class ExtractedAreaDataOutput(
         fun fromExtractAreaEntity(extractedAreaEntity: ExtractedAreaEntity): ExtractedAreaDataOutput {
             return ExtractedAreaDataOutput(
                 inseeCode = extractedAreaEntity.inseeCode,
-                reportings = extractedAreaEntity.reportings.map { ReportingDataOutput.fromReportingDTO(it).id },
-                regulatoryAreas = extractedAreaEntity.regulatoryAreas.map { fromRegulatoryAreaEntity(it).id },
-                amps = extractedAreaEntity.amps.map { fromAMPEntity(it).id },
-                vigilanceAreas =
-                    extractedAreaEntity.vigilanceAreas.mapNotNull {
-                        VigilanceAreasDataOutput.fromVigilanceArea(
-                            it,
-                        ).id
-                    },
+                reportings = extractedAreaEntity.reportings,
+                regulatoryAreas = extractedAreaEntity.regulatoryAreas,
+                amps = extractedAreaEntity.amps,
+                vigilanceAreas = extractedAreaEntity.vigilanceAreas,
             )
         }
     }