--- a/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/adapters/bff/outputs/dashboards/ExtractedAreaDataOutput.kt
+++ b/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/adapters/bff/outputs/dashboards/ExtractedAreaDataOutput.kt
@@ -4,19 +4,19 @@ import fr.gouv.cacem.monitorenv.domain.entities.dashboard.ExtractedAreaEntity
 
 class ExtractedAreaDataOutput(
     val inseeCode: String?,
-    val reportings: List<Int>,
-    val regulatoryAreas: List<Int>,
-    val amps: List<Int>,
-    val vigilanceAreas: List<Int>,
+    val reportingIds: List<Int>,
+    val regulatoryAreaIds: List<Int>,
+    val ampIds: List<Int>,
+    val vigilanceAreaIds: List<Int>,
 ) {
     companion object {
         fun fromExtractAreaEntity(extractedAreaEntity: ExtractedAreaEntity): ExtractedAreaDataOutput {
             return ExtractedAreaDataOutput(
                 inseeCode = extractedAreaEntity.inseeCode,
-                reportings = extractedAreaEntity.reportings,
-                regulatoryAreas = extractedAreaEntity.regulatoryAreas,
-                amps = extractedAreaEntity.amps,
-                vigilanceAreas = extractedAreaEntity.vigilanceAreas,
+                reportingIds = extractedAreaEntity.reportingIds,
+                regulatoryAreaIds = extractedAreaEntity.regulatoryAreaIds,
+                ampIds = extractedAreaEntity.ampIds,
+                vigilanceAreaIds = extractedAreaEntity.vigilanceAreaIds,
             )
         }
     }