--- a/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/endpoints/bff/v1/Dashboards.kt
+++ b/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/endpoints/bff/v1/Dashboards.kt
@@ -14,10 +14,14 @@ import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.websocket.server.PathParam
 import org.locationtech.jts.io.WKTReader
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.*
-import java.util.*
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PutMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestParam
+import org.springframework.web.bind.annotation.RestController
+import java.util.UUID
 
 @RestController
 @RequestMapping("/bff/v1/dashboards")
@@ -60,8 +64,7 @@ class Dashboards(
         @PathParam("Dashboard id")
         @PathVariable(name = "dashboardId")
         dashboardId: UUID,
-    ): ResponseEntity<DashboardDataOutput> {
-        return ResponseEntity.status(HttpStatus.OK)
-            .body(fromDashboardEntity(getDashboard.execute(dashboardId)))
+    ): DashboardDataOutput {
+        return fromDashboardEntity(getDashboard.execute(dashboardId))
     }
 }