--- a/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/JpaReportingRepository.kt
+++ b/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/JpaReportingRepository.kt
@@ -71,6 +71,10 @@ class JpaReportingRepository(
         return dbReportingRepository.findById(reportingId).get().toReportingDTO(mapper)
     }
 
+    override fun findAllById(reportingId: List<Int>): List<ReportingDTO> {
+        return dbReportingRepository.findAllById(reportingId).map { it.toReportingDTO(mapper) }
+    }
+
     @Transactional
     override fun findAll(
         pageNumber: Int?,
@@ -244,9 +248,8 @@ class JpaReportingRepository(
     }
 
     @Transactional
-    override fun findAllByGeometry(geometry: Geometry): List<ReportingDTO> {
-        val reportings = dbReportingRepository.findAllByGeom(geometry = geometry)
-        return reportings.map { it.toReportingDTO(mapper) }
+    override fun findAllIdByGeometry(geometry: Geometry): List<Int> {
+        return dbReportingRepository.findAllByGeom(geometry = geometry)
     }
 
     @Transactional