--- a/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/JpaVigilanceAreaRepository.kt
+++ b/backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/JpaVigilanceAreaRepository.kt
@@ -51,9 +51,8 @@ class JpaVigilanceAreaRepository(
         return dbVigilanceAreaRepository.findAllByIsDeletedIsFalse().map { it.toVigilanceAreaEntity() }
     }
 
-    override fun findAllByGeometry(geometry: Geometry): List<VigilanceAreaEntity> {
-        val vigilanceAreas = dbVigilanceAreaRepository.findAllByGeom(geometry)
-        return vigilanceAreas.map { it.toVigilanceAreaEntity() }
+    override fun findAllIdByGeometry(geometry: Geometry): List<Int> {
+        return dbVigilanceAreaRepository.findAllIdByGeom(geometry)
     }
 
     @Transactional