… backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/dashboard/DashboardEntity.kt
… backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/interfaces/IDBAMPRepository.kt
… backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/interfaces/IDBRegulatoryAreaRepository.kt
… backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/interfaces/IDBReportingRepository.kt
… backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/interfaces/IDBVigilanceAreaRepository.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/domain/use_cases/dashboard/ExtractAreaUTest.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/domain/use_cases/dashboard/fixtures/DashboardFixture.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/domain/use_cases/reportings/GetReportingsByIdsUTest.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/domain/use_cases/reportings/fixtures/ReportingFixture.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/endpoints/bff/v1/DashboardsITests.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/endpoints/bff/v1/ReportingsITests.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/JpaAMPRepositoryTests.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/JpaDashboardsRepositoryITest.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/JpaRegulatoryAreaRepositoryITests.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/JpaReportingRepositoryITests.kt
… backend/src/test/kotlin/fr/gouv/cacem/monitorenv/infrastructure/database/repositories/JpaVigilanceAreaRepositoryITests.kt
… frontend/cypress/e2e/main_window/dashboard/dashboard.spec.ts
… frontend/cypress/e2e/side_window/dashboard/edit_dashboard.spec.ts
… frontend/src/api/api.ts
… frontend/src/api/dashboardsAPI.ts
… frontend/src/api/reportingsAPI.ts
… frontend/src/features/Dashboard/components/DashboardForm/ControlUnits/Item.tsx
… frontend/src/features/Dashboard/components/DashboardForm/ControlUnits/SelectedControlUnits/index.tsx
… frontend/src/features/Dashboard/components/DashboardForm/DashboardForm.tsx
… frontend/src/features/Dashboard/components/DashboardForm/Reportings/Layer.tsx
… frontend/src/features/Dashboard/components/DashboardsList/Columns/index.tsx
… frontend/src/features/Dashboard/components/DashboardsList/Rows/TotalSelectedItemsCell.tsx
… frontend/src/features/Dashboard/components/Layers/index.tsx
… frontend/src/features/Dashboard/slice.ts
… frontend/src/features/Dashboard/types.ts