## What have I done in this PR

It doesnt change dashboard mecanism, only how data are retrieved. The goal was to make payload lighter

- Extract area retrieve only id from geometry not all the datas
- Add a new endpoint for reportings to retrieve reportings from a list of ids
- [x] add tests back I just realized I've forgot them to do quick and dirty refacto.
- [x] add picture from integration env to prove refacto worked

## Before area center is Rennes to Orléans (4.5s server + 5s browser)
<img width="1463" alt="Capture d’écran 2024-10-21 à 11 31 33" src="https://github.com/user-attachments/assets/68f04102-c5a0-4aa0-8ae6-b1ede8ec0de3">

## After  (1s + 1s server + 100ms browser)
<img width="1455" alt="Capture d’écran 2024-10-21 à 11 39 37" src="https://github.com/user-attachments/assets/9512916f-d841-4de2-807f-8dad41eee2c2">

<img width="1443" alt="Capture d’écran 2024-10-21 à 11 39 33" src="https://github.com/user-attachments/assets/75e5e3ce-2d33-4359-8439-f76f9150c0f2">


- [x] Tests E2E (Cypress)